Cleaning of that little mess and extracting Classes to separate files
