+++ Simple Banking System/Simple Banking System/task/banking/BankAccount.py
@@ -1,0 +1,9 @@
+from .BankCard import BankCard
+
+
+class BankAccount:
+    cards = []
+
+    def __init__(self):
+        self.card = BankCard()
+        BankAccount.cards.append(self.card)
